Refactored base class for code generation, so the visitor code is separated for different types. For now using static types, no DI.
